requirements.txt
playwright
google-generativeai
python-dotenv
pillow
easyocr
anticaptchaofficial
fastapi
uvicorn
